Add lab 6 code
Add code for lab 6
