+++ package.json
@@ -1,6 +1,7 @@
 {
   "dependencies": {
     "cookie-parser": "^1.4.6",
+    "cors": "^2.8.5",
     "dotenv": "^16.3.1",
     "express": "^4.18.2",
     "express-oauth2-jwt-bearer": "^1.6.0",
