+++ src/lab5/token_auth/index.html
@@ -155,45 +155,47 @@
                     Authorization: token
                 }
             }).then((response) => {
-                const { access_token, expired, refresh } = response.data;
+                const { username, expired, refresh_token, access_token } = response.data;
 
-                if (access_token) {
-                    const body = {
-                        access_token: access_token,
-                    };
+                if (expired) {
+                    axios.get('/logout');
+                }
 
-                    if (expired) {
-                        axios.get('/logout');
-                    } 
-                    
-                    if (refresh) {
-                        axios.post('/refresh', {
-                            body: body,
-                        }).then((response) => {
-                            location.reload();
-                        }).catch((error) => {});
-                    }
-
+                if (refresh_token) {
                     axios({
-                        method: "get",
-                        url: '/user',
-                        data: { access_token },
+                        method: 'post',
+                        url: '/refresh',
+                        headers: {
+                            "content-type": "application/x-www-form-urlencoded",
+                            Authorization: "Bearer " + access_token,
+                        },
+                        data: {
+                            refresh_token,
+                        },
                     }).then((response) => {
-                        console.log(response.data);
-                        const { username } = response.data;
-                        if (username) {
-                            const mainHolder = document.getElementById("main-holder");
-                    
-                            loginForm.remove();
-                            registerForm.remove();
-                            loginErrorMsg.remove();
-                            loginHeader.remove();
-                            registerHeader.remove();
-
-                            mainHolder.append(`Hello ${username}`);
-                            logoutLink.style.opacity = 1;
-                        }
+                        location.reload();
                     }).catch((error) => {});
+                }
+
+                if (username) {
+                    axios({
+                        method:'post', 
+                        url: "/token/verify", 
+                        headers: { 
+                            "content-type": "application/x-www-form-urlencoded",
+                            Authorization: "Bearer " + access_token,
+                        },
+                    }).then((response) => {
+                        const mainHolder = document.getElementById("main-holder");
+                        loginForm.remove();
+                        registerForm.remove();
+                        loginErrorMsg.remove();
+                        loginHeader.remove();
+                        registerHeader.remove();
+
+                        mainHolder.append(`Hello ${username}`);
+                        logoutLink.style.opacity = 1;
+                    });
                 }
             });
         }
